Extracted annotations to a utility class
(699)
